TODOs: - Finish CSS on ProductDisplay side
