--- client/src/components/Overview/BottomBar.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import FeatureDescription from './BottomBar/FeatureDescription.jsx';
-import SocialMediaButtons from './BottomBar/SocialMediaButtons.jsx';
-
-const BottomBar = () => {
-
-  return(
-    <div className='BottomBar'>
-        <SocialMediaButtons />
-        <FeatureDescription />
-    </div>
-    )
-}
-
-export default BottomBar;
