--- client/src/components/Overview/BottomBar/FeatureContainer.jsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import FeatureAndMaterial from './FeatureAndMaterial.jsx';
-
-const FeatureContainer = (props) => {
-  let features = props.features;
-  return(
-    <div className='FeatureDisplay'>
-      {features.map((detail, index)=>{
-        return (
-          <div key={index}>
-            <FeatureAndMaterial detail={detail} index={index} />
-
-          </div>
-        )
-      })}
-    </div>
-  )
-}
-
-export default FeatureContainer;
