--- client/src/components/Overview/BottomBar/FeatureTitle.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const FeatureTitle = (props) => {
-  return (
-    <div className='FeatureAndMaterialTitle'>
-        <p className='Feature-Spacer'></p>
-        <div key={0} className='FeatureTitle'>Features</div>
-    </div>
-  )
-}
-
-export default FeatureTitle;
