+++ client/src/components/Overview/ProductDetails.jsx
@@ -5,6 +5,8 @@ import StyleInfo from './ProductDetails/StyleInfo.jsx';
 import StyleSelector from './ProductDetails/StyleSelector.jsx';
 import AddToCart from './ProductDetails/AddToCart.jsx';
 import ProductDescription from './ProductDetails/ProductDescription.jsx'
+import FeatureDescription from './ProductDetails/FeatureDescription.jsx';
+import SocialMediaButtons from './ProductDetails/SocialMediaButtons.jsx';
 
 
 const ProductDetails = () => {
@@ -16,6 +18,8 @@ const ProductDetails = () => {
       <StyleSelector />
       <AddToCart />
       <ProductDescription />
+      <FeatureDescription />
+      <SocialMediaButtons />
     </div>
   )
 };
