+++ client/src/styles/Overview/Overview.css
@@ -25,7 +25,7 @@ a:active {
 .Overview-Container {
   width: 82%;
   margin: auto;
-  height: 750px;
+  height: 725px;
   display: flex;
   justify-content: center;
 }
@@ -608,8 +608,7 @@ a:active {
 .QuantityAndCartRow {
   display: flex;
   justify-content: space-around;
-  padding: 20px 0;
-  padding-bottom: 30px;
+  padding: 14px 0 20px 0;
   border-bottom: 1px solid #f5f4f4;
 }
 
@@ -725,43 +724,23 @@ a:active {
 }
 
 .FeatureContainer {
-  border: solid;
-  border-color: rgb(0, 0, 0);
   display: flex;
-  flex-direction: column;
-  justify-content: center;
-  min-width: 20%;
-  width: fit-content;
+  padding: 20px 0 20px 0;
+  border-bottom: 1px solid #f5f4f4;
 }
 
 .FeatureAndMaterialTitle {
   display: flex;
-  align-items: center;
-  height: 4vh;
-}
-
-.Feature-Spacer {
-  min-width: 15%;
+  width: 110px;
+  font-size: 14px;
+  color: #737373;
 }
 
 .FeatureAndMaterial {
   display: flex;
-}
-
-.FeatureTitle {
-  border-bottom: 3px solid black;
-  height: fit-content;
-  font-size: 1.75vh;
-  display: flex;
-  align-items: center;
-}
-
-.FeatureMaterialSpacer {
-  width: 4%;
-}
-
-.FeatureNoMaterial, .MaterialNoFeature {
-  display: flex;
+  justify-content: space-between;
+  width: 274px;
+  padding: 0 0 8px 0;
 }
 
 .Feature {
@@ -770,6 +749,7 @@ a:active {
   justify-content: center;
   align-items: center;
   text-align: center;
+  font-size: 14px;
 }
 
 .Material {
@@ -777,19 +757,14 @@ a:active {
   display: flex;
   align-items: center;
   text-align: left;
-}
-
-.FeatureSpacer {
-  height: 10px;
+  font-size: 14px;
 }
 
 .Social-Media-Buttons {
-  width: 25%;
   display: flex;
   justify-content: space-around;
   align-items: stretch;
-  border: solid;
-  border-color: rgb(0, 0, 0);
+  border-bottom: 1px solid #f5f4f4;
 }
 
 .btn {
@@ -800,28 +775,22 @@ a:active {
 .Info-Container {
   display: flex;
   flex-direction: column;
+  border-bottom: 1px solid #f5f4f4;
+  padding: 20px 0;
 }
 
 .Slogan {
   font-size: 16px;
   font-weight: bold;
-  margin: 20px 0 16px 0;
-}
-
-.SloganContainer {
-  display: flex;
-  height: fit-content;
+  margin: 0 0 16px 0;
 }
 
 .Description {
   font-size: 14px;
   margin: 0;
-  padding: 0 0 12px 0;
-  border-bottom: 1px solid #f5f4f4;
 }
 
-.SMBRow {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
+.SMButton {
+  font-size: 30px;
+  padding: 20px 0;
 }
